# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Social-Media-Api
  version: v1
  description: "Спецификация для проекта Social-Media-Api."
servers:
  - url: http://localhost:8080
    description: local server
paths:
...
  /api/v1/user/createFriendRequest:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Отправка запроса <добавить в друзья>"
      operationId: createFriendRequest
      parameters:
        - name: userIdTo
          description: "ID пользователя, которому отправили заявку <добавить в друзья>"
          in: query
          required: true
          allowEmptyValue: false
          example: 2
          schema:
            type: integer
      responses:
        '200':
          description: Friend request successful post
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 1 send friend request to user with id = 2"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/acceptFriendRequest:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Принятие запроса <добавить в друзья>"
      operationId: acceptFriendRequest
      parameters:
        - name: userIdSendedRequest
          description: "ID пользователя отправляющего заявку <добавить в друзья>"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Friend successful accept friend request
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 2 accept friend request to user with id = 1"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/refuseFriendRequest:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Отклонение запроса <добавить в друзья>"
      operationId: refuseFriendRequest
      parameters:
        - name: userIdSendedRequest
          description: "ID пользователя отправившего заявку <добавить в друзья>"
          in: query
          required: true
          allowEmptyValue: false
          example: 1
          schema:
            type: integer
      responses:
        '200':
          description: Friend request successful refuse
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 2 refused friend request to user with id = 1"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/refuseFollower:
    post:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Отклонение подписки на человека"
      operationId: refuseFollower
      parameters:
        - name: userId
          description: "ID пользователя, на которого подписан пользователь"
          in: query
          required: true
          allowEmptyValue: false
          example: 2
          schema:
            type: integer
      responses:
        '200':
          description: Friend successful refuse follower
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 1 refused follower to user with id = 2"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/deleteFriend:
    delete:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Удаление пользователя из друзей"
      operationId: deleteFriend
      parameters:
        - name: userIdDeleted
          description: "ID пользователя, которого хотят удалить из друзей"
          in: query
          required: true
          allowEmptyValue: false
          example: 2
          schema:
            type: integer
      responses:
        '200':
          description: Friend successful delete friend
          content:
            '*/*':
              schema:
                type: string
                example: "User with id = 1 delete friend with id = 2"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/getAllUsers:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Получение всех пользователей из базы данных (Доступен только авторизованным пользователям с ролью ADMIN)"
      operationId: getAllUsers
      responses:
        '200':
          description: Successful get all users from BD
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/getIdUserByUsername:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Получение id пользователя по его имени"
      operationId: getIdUserByUsername
      parameters:
        - name: username
          description: "Имя пользователя, по которому ищется userId"
          in: query
          required: true
          allowEmptyValue: false
          example: Sergey
          schema:
            type: string
      responses:
        '200':
          description: Successful get id users
          content:
            '*/*':
              schema:
                type: string
                example: "User id = 1 by username - Sergey"
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/getAllUsersFriends:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Получение всех друзей пользователя"
      operationId: getAllUsersFriends
      responses:
        '200':
          description: Successful get all friends user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/user/getAllUsersFollowers:
    get:
      security:
        - bearerAuth: [ ]
      tags:
        - User endpoints
      summary: "Получение всех друзей фолловеров"
      operationId: getAllUsersFollowers
      responses:
        '200':
          description: Successful get all followers user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Not forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserDto:
      required:
        - id
        - email
        - name
        - password
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: "Id пользователя"
          example: "1"
        name:
          type: string
          description: "Имя пользователя"
          example: "Sergey"
        password:
          type: string
          description: "Пароль пользователя"
          example: "111"
          writeOnly: true
        email:
          type: string
          description: "Электронная почта пользователя"
          example: "Sergey@gmail.com"
    ErrorResponse:
      type: object
      properties:
        messageError:
          type: string
          description: Error message
          example: "Error"